V0 of the refactored libCEED.
